Excepciones

+++ Clinica-MVC/src/main/java/dh/backend/clinicamvc/service/impl/PacienteService.java
@@ -1,6 +1,7 @@
 package dh.backend.clinicamvc.service.impl;
 
 import dh.backend.clinicamvc.entity.Paciente;
+import dh.backend.clinicamvc.exception.ResourceNotFoundException;
 import dh.backend.clinicamvc.repository.IPacienteRepository;
 import dh.backend.clinicamvc.service.IPacienteService;
 import org.springframework.stereotype.Service;
@@ -35,7 +36,11 @@ public class PacienteService implements IPacienteService {
     }
 
     @Override
-    public void eliminarPaciente(Integer id) {
+    public void eliminarPaciente(Integer id) throws ResourceNotFoundException {
+        Optional<Paciente> pacienteOptional = buscarPorId(id);
+        if(pacienteOptional.isPresent())
         pacienteRepository.deleteById(id);
+        else
+            throw new ResourceNotFoundException("{\"mensaje\":\"paciente no encontrado\"}");
     }
 }
